moved login use_cases into one package for organization sakes
