+++ src/main/java/use_case/login/club_login/ClubLoginOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.club_login;
+package use_case.login.club_login;
 
 /**
  * Output boundary for the login use case for clubs.
