+++ src/main/java/use_case/login/student_login/StudentLoginOutputBoundary.java
@@ -1,4 +1,4 @@
-package use_case.student_login;
+package use_case.login.student_login;
 
 /**
  * The output boundary for the student Login Use Case.
